move createPool to own proto package
