+++ x/concentrated-liquidity/types/codec.go
@@ -1,4 +1,4 @@
-package model
+package types
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -6,22 +6,22 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/msgservice"
 	authzcodec "github.com/cosmos/cosmos-sdk/x/authz/codec"
-
-	"github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity/types"
 )
 
 func RegisterCodec(cdc *codec.LegacyAmino) {
-	cdc.RegisterInterface((*types.ConcentratedPoolExtension)(nil), nil)
-	cdc.RegisterConcrete(&MsgCreateConcentratedPool{}, "osmosis/create-concentrated-pool", nil)
+	cdc.RegisterInterface((*ConcentratedPoolExtension)(nil), nil)
+	cdc.RegisterConcrete(&MsgCreatePosition{}, "osmosis/cl-create-position", nil)
+	cdc.RegisterConcrete(&MsgWithdrawPosition{}, "osmosis/cl-withdraw-position", nil)
 }
 
 func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 	registry.RegisterImplementations(
 		(*sdk.Msg)(nil),
-		&MsgCreateConcentratedPool{},
+		&MsgCreatePosition{},
+		&MsgWithdrawPosition{},
 	)
 
-	msgservice.RegisterMsgServiceDesc(registry, &_MsgCreator_serviceDesc)
+	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)
 }
 
 var (
